+++ native/src/main/java/module-info.java
@@ -22,6 +22,7 @@ module io.ballerina.lib.sap {
     requires io.ballerina.runtime;
     requires org.slf4j;
     requires io.ballerina.stdlib.time;
+    requires java.xml;
     exports io.ballerina.lib.sap;
     exports io.ballerina.lib.sap.parameterprocessor;
     exports io.ballerina.lib.sap.dataproviders;